+++ packages/table/src/helpers/filters/booleanTextFilter.ts
@@ -1,6 +1,6 @@
-import { TBooleanFilterValue } from '../../filters';
-
-export const booleanTextFilter = (data: boolean, { value }: TBooleanFilterValue) =>
-  value === '' ||
-  (value === 'true' && data === true) ||
-  (value === 'false' && data === false);
+export const booleanTextFilter = (data: boolean, search: string) =>
+  search === '' ||
+  (search.toLowerCase() === 'true' && data === true) ||
+  (search.toLowerCase() === 'false' && data === false) ||
+  (search.toLowerCase() === 'yes' && data === true) ||
+  (search.toLowerCase() === 'no' && data === false);